% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{gpt_neox_from_pretrained}
\alias{gpt_neox_from_pretrained}
\alias{gpt_neox_from_config}
\title{GPTNeoX model}
\usage{
gpt_neox_from_pretrained(identifier, ..., revision = "main")

gpt_neox_from_config(config)
}
\arguments{
\item{identifier}{The HuggingFace repository to download the model from.}

\item{...}{Not currently used.}

\item{revision}{The repository revision to download from. Either a branch name
or a commit hash.}

\item{config}{A model configuration created by \code{\link[=gpt_neox_config]{gpt_neox_config()}} or obtained
from \code{\link[=gpt_neox_config_from_pretrained]{gpt_neox_config_from_pretrained()}}}
}
\description{
Creates a GPTNeoX Model.
}
\section{Functions}{
\itemize{
\item \code{gpt_neox_from_pretrained()}: Creates from a configuration from a HF repository.

\item \code{gpt_neox_from_config()}: Creates a GPTNeoX model from a configuration list.

}}
