% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generation.R
\name{gpt_neox_generate}
\alias{gpt_neox_generate}
\title{Samples tokens from GPT NeoX model}
\usage{
gpt_neox_generate(
  model,
  tokenizer,
  prompt,
  ...,
  config = list(),
  verbose = TRUE
)
}
\arguments{
\item{model}{A GPTNeoX Model as created with \code{\link[=gpt_neox_from_pretrained]{gpt_neox_from_pretrained()}}}

\item{tokenizer}{A GPTNeoX tokenizer as created with \code{\link[=gpt_neox_tokenizer_from_pretrained]{gpt_neox_tokenizer_from_pretrained()}}}

\item{prompt}{A text prompt.}

\item{config}{Configuration for generation.}

\item{verbose}{Wether to print output during generation.}
}
\description{
Samples tokens from GPT NeoX model
}
