% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{gpt_neox_tokenizer_from_pretrained}
\alias{gpt_neox_tokenizer_from_pretrained}
\title{Creates a GPTNeoX tokenizer from a pre-trained tokenizer}
\usage{
gpt_neox_tokenizer_from_pretrained(identifier, ..., revision = "main")
}
\arguments{
\item{identifier}{The HuggingFace repository to download the model from.}

\item{...}{Not currently used.}

\item{revision}{The repository revision to download from. Either a branch name
or a commit hash.}
}
\description{
Creates a GPTNeoX tokenizer from a pre-trained tokenizer
}
