% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gptneox.R
\name{GPTNeoXAttention}
\alias{GPTNeoXAttention}
\alias{GPTNeoXMLP}
\alias{GPTNeoXLayer}
\alias{GPTNeoXModel}
\alias{GPTNeoXForCausalLM}
\title{GPTNeoX modules}
\usage{
GPTNeoXAttention(config)

GPTNeoXMLP(config)

GPTNeoXLayer(config)

GPTNeoXModel(config)

GPTNeoXForCausalLM(config)
}
\arguments{
\item{config}{A model configuration created by \code{\link[=gpt_neox_config]{gpt_neox_config()}} or obtained
from \code{\link[=gpt_neox_config_from_pretrained]{gpt_neox_config_from_pretrained()}}}
}
\description{
GPTNeoX modules
}
\section{Functions}{
\itemize{
\item \code{GPTNeoXAttention()}: Attention module

\item \code{GPTNeoXMLP()}: MLP layer module

\item \code{GPTNeoXLayer()}: Layer module

\item \code{GPTNeoXModel()}: Model module

\item \code{GPTNeoXForCausalLM()}: CausalLM

}}
